+++ tomcat/src/main/java/org/apache/http/HttpResponse.java
@@ -9,6 +9,9 @@ import java.util.List;
 import java.util.Map;
 import org.apache.exception.RequestProcessingException;
 import org.apache.exception.SocketWriteException;
+import org.apache.http.value.HttpHeader;
+import org.apache.http.value.HttpVersion;
+import org.apache.http.value.StatusCode;
 
 public class HttpResponse {
