+++ tomcat/src/main/java/org/apache/http/value/ContentType.java
@@ -1,4 +1,4 @@
-package org.apache.http;
+package org.apache.http.value;
 
 import java.util.List;
