+++ tomcat/src/main/java/org/apache/http/value/HttpMethod.java
@@ -1,4 +1,4 @@
-package org.apache.http;
+package org.apache.http.value;
 
 public enum HttpMethod {
     GET, POST, PUT, DELETE, PATCH, HEAD, OPTIONS, CONNECT
