+++ tomcat/src/main/java/org/apache/http/value/HttpVersion.java
@@ -1,4 +1,4 @@
-package org.apache.http;
+package org.apache.http.value;
 
 import java.util.List;
